feat: Move some of the remaining items around
